backend/requirements.txt
# NeuralEdge AI SaaS Backend
# FastAPI + PostgreSQL + Redis + Celery

# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# === Database ===
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# === Cache / Message Queue ===
redis[hiredis]==5.2.1
celery[redis]==5.4.0

# === Security ===
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
cryptography==44.0.0
python-multipart==0.0.18

# === HTTP Client ===
httpx==0.28.1

# === WebSocket ===
websockets==14.1

# === Utilities ===
python-dotenv==1.0.1
email-validator==2.2.0
orjson==3.10.14

# === Exchange ===
ccxt==4.4.54

# === Monitoring ===
prometheus-client==0.21.1
